Rename package name to "super_enum" from "sealed"
