--- sealed/lib/sealed.dart
@@ -1,5 +1,0 @@
-library sealed;
-
-export 'package:meta/meta.dart' hide sealed;
-export 'src/annotations.dart'
-    show sealed, generic, Generic, Sealed, object, Object, DataField, Data;
