--- sealed_generator/lib/sealed_generator.dart
@@ -1,7 +1,0 @@
-library sealed_generator;
-
-import 'package:build/build.dart';
-import 'src/index.dart';
-
-Builder sealedGeneratorFactory(BuilderOptions options) =>
-    sealedGeneratorFactoryBuilder();
