+++ super_enum/lib/super_enum.dart
@@ -1,0 +1,13 @@
+library super_enum;
+
+export 'package:meta/meta.dart';
+export 'src/annotations.dart'
+    show
+        superEnum,
+        generic,
+        Generic,
+        SuperEnum,
+        object,
+        Object,
+        DataField,
+        Data;
