+++ super_enum_generator/lib/src/class_generator.dart
@@ -1,10 +1,10 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:code_builder/code_builder.dart';
-import 'package:sealed/sealed.dart';
-import 'package:sealed_generator/src/type_processor.dart' as TypeProcessor;
-import 'package:sealed_generator/src/extension.dart';
+import 'package:super_enum/super_enum.dart';
+import 'package:super_enum_generator/src/type_processor.dart' as TypeProcessor;
+import 'package:super_enum_generator/src/extension.dart';
 import 'package:dart_style/dart_style.dart';
-import 'package:sealed_generator/src/references.dart' as References;
+import 'package:super_enum_generator/src/references.dart' as References;
 
 class ClassGenerator {
   final ClassElement element;
