+++ super_enum_generator/lib/src/extension.dart
