+++ super_enum_generator/lib/src/index.dart
