+++ super_enum_generator/lib/src/references.dart
