+++ super_enum_generator/lib/super_enum_generator.dart
@@ -1,0 +1,7 @@
+library super_enum_generator;
+
+import 'package:build/build.dart';
+import 'src/index.dart';
+
+Builder superEnumGeneratorFactory(BuilderOptions options) =>
+    superEnumGeneratorFactoryBuilder();
