+++ super_enum_generator/test/super_enum_generator_test.dart
@@ -1,10 +1,10 @@
-import 'package:sealed/sealed.dart';
-import 'package:sealed_generator/src/generator.dart';
+import 'package:super_enum/super_enum.dart';
+import 'package:super_enum_generator/src/generator.dart';
 import 'package:source_gen_test/source_gen_test.dart';
 
 Future<void> main() async {
   final reader = await initializeLibraryReaderForDirectory(
       'test/src', 'generator_test_src.dart');
   initializeBuildLogTracking();
-  testAnnotatedElements<Sealed>(reader, SealedGenerator());
+  testAnnotatedElements<SuperEnum>(reader, SuperEnumGenerator());
 }
